chore: move prefix checks together with other SAP checks
