+++ sn_node/src/node/handover/handover_consensus.rs
@@ -3,7 +3,6 @@ use core::fmt::Debug;
 use std::cmp::Ordering;
 use std::collections::BTreeMap;
 use tracing::info;
-use xor_name::Prefix;
 
 use sn_consensus::consensus::{Consensus, VoteResponse};
 use sn_consensus::vote::{Ballot, SignedVote, Vote};
@@ -18,7 +17,6 @@ use sn_interface::network_knowledge::SapCandidate;
 #[derive(Debug, Clone)]
 pub(crate) struct Handover {
     pub(crate) consensus: Consensus<SapCandidate>,
-    pub(crate) section_prefix: Prefix,
     pub(crate) failed_consensus_rounds:
         BTreeMap<Generation, (Consensus<SapCandidate>, Decision<SapCandidate>)>,
     /// Handover gen starting at 0, and then +1 for each retries after failed consensus rounds
@@ -30,11 +28,9 @@ impl Handover {
         secret_key: (NodeId, SecretKeyShare),
         elders: PublicKeySet,
         n_elders: usize,
-        section_prefix: Prefix,
     ) -> Self {
         Handover {
             consensus: Consensus::<SapCandidate>::from(secret_key, elders, n_elders),
-            section_prefix,
             failed_consensus_rounds: BTreeMap::new(),
             gen: 0,
         }
@@ -47,7 +43,6 @@ impl Handover {
             faults: self.consensus.faults(),
         };
         let signed_vote = self.sign_vote(vote)?;
-        self.validate_proposals(&signed_vote)?;
         signed_vote
             .detect_byzantine_faults(
                 &self.consensus.elders,
@@ -144,7 +139,6 @@ impl Handover {
             self.id(),
             signed_vote
         );
-        self.validate_proposals(&signed_vote)?;
 
         match signed_vote.vote.gen.cmp(&self.gen) {
             Ordering::Less => self.handle_outdated_signed_vote(signed_vote),
@@ -173,42 +167,6 @@ impl Handover {
             None
         }
     }
-
-    pub(crate) fn validate_proposals(&self, signed_vote: &SignedVote<SapCandidate>) -> Result<()> {
-        signed_vote
-            .proposals()
-            .into_iter()
-            .try_for_each(|prop| self.validate_proposal(prop))
-    }
-
-    pub(crate) fn validate_proposal(&self, proposal: SapCandidate) -> Result<()> {
-        match proposal {
-            SapCandidate::ElderHandover(single_sap) => {
-                // single handover, must be same prefix
-                if single_sap.prefix() == self.section_prefix {
-                    Ok(())
-                } else {
-                    Err(Error::InvalidSectionPrefixForCandidate)
-                }
-            }
-            SapCandidate::SectionSplit(sap1, sap2) => {
-                // section split, must be 2 distinct children prefixes
-                let our_p = &self.section_prefix;
-                let p1 = sap1.prefix();
-                let p2 = sap2.prefix();
-                if p1.is_extension_of(our_p)
-                    && p2.is_extension_of(our_p)
-                    && p1.bit_count() == our_p.bit_count() + 1
-                    && p2.bit_count() == our_p.bit_count() + 1
-                    && p1 != p2
-                {
-                    Ok(())
-                } else {
-                    Err(Error::InvalidSectionPrefixForSplitCandidates)
-                }
-            }
-        }
-    }
 }
 
 #[cfg(test)]
@@ -229,7 +187,6 @@ mod tests {
                 (id, elders_sk.secret_key_share(id as usize)),
                 elders_sk.public_keys(),
                 7,
-                Prefix::default(),
             )
         }));
